rename `Arrow` -> `ArrowSection`
The `Arrow` component will in the future be a group of one or more ArrowSections
